Clean up. Fixed bug where elements could not be embedded into a pool.
