+++ src/components/nodes/poolLane/poolLane.vue
@@ -25,7 +25,6 @@ import pull from 'lodash/pull';
 import { poolColor } from '@/components/nodeColors';
 import CrownConfig from '@/components/crownConfig';
 import highlightConfig from '@/mixins/highlightConfig';
-import poolConfig from '@/mixins/poolConfig';
 
 export default {
   components: {
@@ -45,7 +44,7 @@ export default {
     'planeElements',
     'isRendering',
   ],
-  mixins: [highlightConfig, resizeConfig, poolConfig],
+  mixins: [highlightConfig, resizeConfig],
   data() {
     return {
       shape: null,
@@ -55,29 +54,6 @@ export default {
   watch: {
     'node.definition.name'(name) {
       this.shape.attr('label/text', name);
-    },
-  },
-  methods: {
-    // @todo check this
-    removeShape() {
-      this.$emit('remove-node', this.node);
-
-      const poolComponent = this.node.pool.component;
-      const sortedLanes = poolComponent.sortedLanes();
-
-      if (sortedLanes.length === 2) {
-      /* Do not allow pool with only one lane;
-       * if removing 2nd last lane, remove the other lane as well */
-        this.$emit('remove-node', sortedLanes.filter(lane => lane !== this.shape)[0].component.node);
-        return;
-      }
-
-      if (this.shape === sortedLanes[sortedLanes.length - 1]) {
-        poolComponent.fillLanes(this.shape, 'top-right', true);
-        return;
-      }
-
-      poolComponent.fillLanes(this.shape, 'bottom-right', true);
     },
   },
   mounted() {
